+++ src/TestableHttpClient/Utils/StringMatcher.cs
@@ -4,7 +4,7 @@ namespace TestableHttpClient.Utils;
 
 internal static class StringMatcher
 {
-    internal static bool Matches(string value, string pattern)
+    internal static bool Matches(string value, string pattern, bool ignoreCase = false)
     {
         var escapedPattern = Regex.Escape(pattern);
 #if NETSTANDARD2_0
@@ -12,7 +12,7 @@ internal static class StringMatcher
 #else
         var regex = escapedPattern.Replace("\\*", "(.*)", StringComparison.InvariantCultureIgnoreCase);
 #endif
-
-        return Regex.IsMatch(value, $"^{regex}$");
+        RegexOptions options = ignoreCase ? RegexOptions.IgnoreCase : RegexOptions.None;
+        return Regex.IsMatch(value, $"^{regex}$", options);
     }
 }