+++ test/TestableHttpClient.Tests/HttpRequestMessagesCheckExtensionsTests/WithRequestUri.cs
@@ -4,11 +4,10 @@ namespace TestableHttpClient.Tests.HttpRequestMessagesCheckExtensionsTests;
 
 public class WithRequestUri
 {
-#nullable disable
     [Fact]
     public void WithRequestUri_WithoutNumberOfRequests_NullCheck_ThrowsArgumentNullException()
     {
-        IHttpRequestMessagesCheck sut = null;
+        IHttpRequestMessagesCheck sut = null!;
 
         var exception = Assert.Throws<ArgumentNullException>(() => sut.WithRequestUri("*"));
 
@@ -18,7 +17,7 @@ public class WithRequestUri
     [Fact]
     public void WithRequestUri_WithNumberOfRequests_NullCheck_ThrowsArgumentNullException()
     {
-        IHttpRequestMessagesCheck sut = null;
+        IHttpRequestMessagesCheck sut = null!;
 
         var exception = Assert.Throws<ArgumentNullException>(() => sut.WithRequestUri("*", 2));
 
@@ -37,7 +36,6 @@ public class WithRequestUri
         Assert.Equal("pattern", exception.ParamName);
         sut.Verify(x => x.WithFilter(Its.AnyPredicate(), It.IsAny<int?>(), It.IsAny<string>()), Times.Never());
     }
-#nullable restore
 
     [Fact]
     public void WithRequestUri_WithoutNumberOfRequests_CallsWithCorrectly()
@@ -50,11 +48,21 @@ public class WithRequestUri
     }
 
     [Fact]
-    public void WithRequestUri_WithNumberOfRequests_CallsWithCorrectly()
+    public void WithRequestUri_WithoutNumberOfRequestsAndNotIgnoringCase_CallsWithCorrectly()
     {
         var sut = new Mock<IHttpRequestMessagesCheck>();
 
-        sut.Object.WithRequestUri("https://example.com/", 2);
+        sut.Object.WithRequestUri("https://example.com/", ignoreCase: false);
+
+        sut.Verify(x => x.WithFilter(Its.AnyPredicate(), null, "uri pattern 'https://example.com/'"));
+    }
+
+    [Fact]
+    public void WithRequestUri_WithNumberOfRequestsAndNotIgnoringCase_CallsWithCorrectly()
+    {
+        var sut = new Mock<IHttpRequestMessagesCheck>();
+
+        sut.Object.WithRequestUri("https://example.com/", ignoreCase: false, 2);
 
         sut.Verify(x => x.WithFilter(Its.AnyPredicate(), (int?)2, "uri pattern 'https://example.com/'"));
     }