+++ CHANGELOG.md
@@ -23,6 +23,7 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - When `HttpResponseMessage.Content` is null after `IResponse.ExecuteAsync` was called, an empty `StringContent` is added (Up until .NET 6.0, since Content is always filled there).
 - The `HttpRequestMessage` is always added to the response, which is now possible, since we no longer allow reusing responses.
 - Added `ConfigureAwait(false)` to all calls, since we now use async/await in the library.
+- The check on request uri is now case insensitive by default, when test url's that are case sensitive set the `ignoreCase` parameter to false.
 
 ## [0.7] - 2022-09-22
 ### Changed